src/test/java/org/example/ServiceTests.java
package org.example;

import static org.assertj.core.api.Assertions.assertThat;
import static org.assertj.core.api.Assertions.assertThatThrownBy;

import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.AfterAll;
import org.junit.jupiter.api.extension.ExtendWith;
import static org.junit.jupiter.api.Assertions.assertThrows;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.BeforeAll;

import org.mockito.junit.jupiter.MockitoExtension;
import static org.mockito.Mockito.mock;
import org.mockito.Mock;
import org.mockito.Mockito;

import org.junit.jupiter.api.Test;
import static org.junit.jupiter.api.Assertions.assertTrue;
import static org.junit.jupiter.api.Assertions.assertFalse;

public class ServiceTests {

    @Test
    void testIsEven() {
        Service service = new Service();
        assertTrue(service.isEven(2));
        assertFalse(service.isEven(3));
    }

    @Test
    void testHighComplexityMethod() {
        Service service = new Service();
        // Test case where a is positive, b is positive, and c is positive
        service.highComplexityMethod(1, 2, 3);
        // Add assertions here for the expected output
        
        // Test case where a is non-positive, b is non-positive, and c is non-positive
        service.highComplexityMethod(-1, -2, -3);
        // Add assertions here for the expected output
        
        // Add more test cases for different combinations of a, b, and c values
    }
}
